+++ test/test_axis_dict.py
@@ -13,6 +13,29 @@ def test_sum():
     assert a1 + b2 == AxisDict({a: 1, b: 2})
     assert a1 + a2 == AxisDict({a: 3})
     assert a1 + 1 == AxisDict({a: 1, NO_AXIS: 1})
+    assert 1 + a1 == AxisDict({a: 1, NO_AXIS: 1})
+
+
+def test_div():
+    a, b = symbols("a b")
+
+    ad = AxisDict({a: 2, b: 4})
+
+    assert ad / 2 == AxisDict({a: 1, b: 2})
+    assert 8 / ad == AxisDict({a: 4, b: 2})
+
+
+def test_sub():
+    a, b = symbols("a b")
+
+    a1 = AxisDict({a: 1})
+    a2 = AxisDict({a: 2})
+    b2 = AxisDict({b: 2})
+
+    assert a1 - b2 == AxisDict({a: 1, b: -2})
+    assert a2 - a1 == AxisDict({a: 1})
+    assert 1 - a1 == AxisDict({a: -1, NO_AXIS: 1})
+    assert a1 - 1 == AxisDict({a: 1, NO_AXIS: -1})
 
 
 def test_mul():
@@ -21,6 +44,17 @@ def test_mul():
     a1 = AxisDict({a: 2})
     assert a1 * 2 == AxisDict({a: 4})
     assert simplify(a1 / 2) == AxisDict({a: 1})
+
+
+def test_equals():
+    a = symbols("a")
+
+    a1 = AxisDict({a: 1})
+    a1_bis = AxisDict({a: 1})
+    a2 = AxisDict({a: 2})
+
+    assert a1.equals(a1_bis)
+    assert not a1.equals(a2)
 
 
 def test_free_symbols():